refactor: use common component for llm search
